adding more logging and error catching

+++ mondupe/lib/mondupe.rb
@@ -115,14 +115,21 @@ class Mondupe
     puts "#{Time.now.to_s} - Dropping existing database"
     `ssh -i #{ssh_key} #{ssh_user}@#{instance_ip} "echo 'db.dropDatabase()' | mongo cde_production"`
     if $?.success? then puts "#{Time.now.to_s} - Database drop complete" else die("Error dropping database") end
+    puts "Extracting database dump archive file..."
+    `ssh -i #{ssh_key} #{ssh_user}@#{instance_ip} "cd #{dump_tmp_path}; tar xf #{dump_file_name}"`
+    if $?.success? then puts "#{Time.now.to_s} - Extraction complete!" else die("Error extracting archive") end
     puts "Restoring Mongo Database from extracted dump: #{File.join(dump_tmp_path, "cde_production")}"
-    `ssh -i #{ssh_key} #{ssh_user}@#{instance_ip} "cd #{dump_tmp_path}; tar xf #{dump_file_name}; time mongorestore /tmp/cde_production"`
+    `ssh -i #{ssh_key} #{ssh_user}@#{instance_ip} "time mongorestore #{File.join(dump_tmp_path, "cde_production")}"`
+    if $?.success? then puts "#{Time.now.to_s} - Database restore complete!" else die("Error restoring databse") end
     puts "Removing database archive file"
     `ssh -i #{ssh_key} #{ssh_user}@#{instance_ip} "rm -rf #{File.join(dump_tmp_path, dump_file_name)}"`
+    if $?.success? then puts "#{Time.now.to_s} - Archive removed!" else die("Error removing archive") end
     puts "#{Time.now.to_s} - Removing saved searches"
     `ssh -i #{ssh_key} #{ssh_user}@#{instance_ip} "mongo cde_production --eval \\"db.users.update({save_searches: {$ne: null}}, {$unset: {save_searches: ''}}, {multi: true})\\""`
+    if $?.success? then puts "#{Time.now.to_s} - Saved searches removed" else die("Error removing saved searches") end
     puts "#{Time.now.to_s} - Cleaning up our mess..."
     `ssh -i #{ssh_key} #{ssh_user}@#{instance_ip} "rm -rf #{File.join(dump_tmp_path, 'cde_production')}"`
+    if $?.success? then puts "#{Time.now.to_s} - Mess cleaned up!" else die("Error cleaning up after myself...") end
   end
 
   def terminate(instance_id)
